+++ CMakeLists.txt
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 
-cmake_minimum_required(VERSION 3.25 FATAL_ERROR)
+cmake_minimum_required(VERSION 3.24 FATAL_ERROR)
 
 
 list(APPEND CMAKE_MESSAGE_CONTEXT "morpheus")
@@ -98,8 +98,6 @@ rapids_cmake_build_type(Release)
 project(morpheus
     VERSION 23.07.00
     LANGUAGES C CXX CUDA)
-
-set(CMAKE_LIBRARY_ARCHITECTURE x86_64-linux-gnu)
 
 rapids_cmake_write_version_file(${CMAKE_BINARY_DIR}/autogenerated/include/morpheus/version.hpp)
