+++ morpheus/_lib/cmake/libmorpheus.cmake
@@ -14,55 +14,10 @@
 
 include(GenerateExportHeader)
 
-# # Place the two cuda sources in their own target and disable IWYU for that target.
-# add_library(cuda_objs
-#   OBJECT
-#     src/utilities/matx_util.cu
-# )
-
-# set_target_properties(
-#   cuda_objs
-#   PROPERTIES
-#     CUDA_STANDARD 17
-#     CUDA_STANDARD_REQUIRED ON
-#     C_INCLUDE_WHAT_YOU_USE ""
-#     CXX_INCLUDE_WHAT_YOU_USE ""
-#     EXPORT_COMPILE_COMMANDS OFF
-# )
-
-# # Set the includes as private because we want the public to be on libmorpheus
-# target_include_directories(cuda_objs
-#   PRIVATE
-#     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-#     $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/autogenerated/include> # Include the generated version file
-#     $<INSTALL_INTERFACE:include>
-# )
-
-# # Set the links as private because we want the public to be on libmorpheus
-# target_link_libraries(cuda_objs
-#   PRIVATE
-#     mrc::pymrc
-#     cudf::cudf
-#     matx::matx
-# )
-
-set(DOCA_BUILD_FILES "")
-# if(MORPHEUS_SUPPORT_DOCA)
-#   set(DOCA_BUILD_FILES
-#     src/doca/doca_context.cpp
-#     src/doca/doca_rx_pipe.cpp
-#     src/doca/doca_rx_queue.cpp
-#     src/doca/doca_semaphore.cpp
-#     src/doca/rte_context.cpp
-#     src/stages/doca_source.cpp
-#     src/stages/doca_source_kernels.cu
-#   )
-# endif()
-
 add_library(morpheus
   # Keep these sorted!
-  src/io/data_loader.cpp
   src/io/data_loader_registry.cpp
+  src/io/data_loader.cpp
   src/io/deserializers.cpp
   src/io/loaders/file.cpp
   src/io/loaders/grpc.cpp
@@ -71,9 +26,9 @@ add_library(morpheus
   src/io/loaders/rest.cpp
   src/io/serializers.cpp
   src/messages/control.cpp
-  src/messages/memory/inference_memory.cpp
   src/messages/memory/inference_memory_fil.cpp
   src/messages/memory/inference_memory_nlp.cpp
+  src/messages/memory/inference_memory.cpp
   src/messages/memory/inference_memory.cpp
   src/messages/memory/response_memory_probs.cpp
   src/messages/memory/response_memory.cpp
@@ -85,8 +40,8 @@ add_library(morpheus
   src/messages/multi_response_probs.cpp
   src/messages/multi_response.cpp
   src/messages/multi_tensor.cpp
-  src/modules/data_loader_module.cpp
   src/messages/multi.cpp
+  src/modules/data_loader_module.cpp
   src/objects/data_table.cpp
   src/objects/dev_mem_info.cpp
   src/objects/dtype.cpp
@@ -114,12 +69,11 @@ add_library(morpheus
   src/stages/write_to_file.cpp
   src/utilities/cudf_util.cpp
   src/utilities/cupy_util.cpp
+  src/utilities/matx_util.cu
   src/utilities/python_util.cpp
   src/utilities/string_util.cpp
   src/utilities/table_util.cpp
   src/utilities/tensor_util.cpp
-  ${DOCA_BUILD_FILES}
-  src/utilities/matx_util.cu
 )
 
 add_library(${PROJECT_NAME}::morpheus ALIAS morpheus)
@@ -142,21 +96,6 @@ target_include_directories(morpheus
     $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/autogenerated/include> # Include the generated version file
     $<INSTALL_INTERFACE:include>
 )
-
-# if(MORPHEUS_SUPPORT_DOCA)
-#   find_package(doca REQUIRED)
-#   target_link_libraries(morpheus
-#     PUBLIC
-#       doca::doca_eth
-#       doca::doca
-#   )
-#   # target_include_directories(morpheus
-#   #     PUBLIC
-#   #         /opt/mellanox/dpdk/include/dpdk
-#   #         /opt/mellanox/dpdk/include/x86_64-linux-gnu/dpdk
-#   #         /usr/local/include
-#   # )
-# endif()
 
 # Generates an include file for specifying external linkage since everything is hidden by default
 generate_export_header(morpheus
