+++ morpheus/_lib/doca/CMakeLists.txt
@@ -12,72 +12,45 @@
 # the License.
 # =============================================================================
 
-include(GenerateExportHeader)
-
-# # Place the two cuda sources in their own target and disable IWYU for that target.
-# add_library(cuda_objs
-#   OBJECT
-#     src/utilities/matx_util.cu
-# )
-
-# set_target_properties(
-#   cuda_objs
-#   PROPERTIES
-#     CUDA_STANDARD 17
-#     CUDA_STANDARD_REQUIRED ON
-#     C_INCLUDE_WHAT_YOU_USE ""
-#     CXX_INCLUDE_WHAT_YOU_USE ""
-#     EXPORT_COMPILE_COMMANDS OFF
-# )
-
-# # Set the includes as private because we want the public to be on libmorpheus
-# target_include_directories(cuda_objs
-#   PRIVATE
-#     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-#     $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/autogenerated/include> # Include the generated version file
-#     $<INSTALL_INTERFACE:include>
-# )
-
-# # Set the links as private because we want the public to be on libmorpheus
-# target_link_libraries(cuda_objs
-#   PRIVATE
-#     mrc::pymrc
-#     cudf::cudf
-#     matx::matx
-# )
-
+# Need to load this first before calling find_package(doca REQUIRED)
 morpheus_configure_libmd()
-
 morpheus_configure_libbsd()
 
 set(doca_ROOT "/opt/mellanox/doca")
-
 find_package(doca REQUIRED)
 
 add_library(morpheus_doca
   # Keep these sorted!
-  doca_context.cpp
-  doca_rx_pipe.cpp
-  doca_rx_queue.cpp
-  doca_semaphore.cpp
-  rte_context.cpp
-  # src/stages/doca_source.cpp
-  doca_source_kernels.cu
+  src/doca_context.cpp
+  src/doca_rx_pipe.cpp
+  src/doca_rx_queue.cpp
+  src/doca_semaphore.cpp
+  src/doca_source_kernels.cu
+  src/doca_source.cpp
+  src/rte_context.cpp
 )
 
 add_library(${PROJECT_NAME}::morpheus_doca ALIAS morpheus_doca)
 
+target_include_directories(morpheus_doca
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:include>
+)
+
 target_link_libraries(morpheus_doca
   PRIVATE
-    ${PROJECT_NAME}::morpheus
     doca::doca
+  PUBLIC
+    ${PROJECT_NAME}::morpheus
+
 )
 
 # Ideally, we dont use glob here. But there is no good way to guarantee you dont miss anything like *.cpp
 file(GLOB_RECURSE morpheus_doca_public_headers
   LIST_DIRECTORIES FALSE
   CONFIGURE_DEPENDS
-  "${CMAKE_CURRENT_SOURCE_DIR}/*.hpp"
+  "${CMAKE_CURRENT_SOURCE_DIR}/include/*.hpp"
 )
 
 # Add headers to target sources file_set so they can be installed
