+++ morpheus/_lib/doca/include/doca_source.hpp
@@ -17,7 +17,8 @@
 
 #pragma once
 
-#include <morpheus/messages/meta.hpp>
+#include "morpheus/messages/meta.hpp"
+
 #include <pymrc/node.hpp>
 
 #include <memory>
